feat: Add Deno build task for Deno Deploy
This commit refactors the project to support a standard Fresh v2 deployment workflow using Vite, making it suitable for Deno Deploy.

- The original `main.ts` has been renamed to `desktop.ts` to preserve the desktop webview functionality.
- A new `main.ts` has been created as a standard Fresh v2 server entrypoint.
- A `vite.config.ts` file has been added to configure the Fresh Vite plugin.
- The `deno.jsonc` tasks have been updated:
  - The `desktop` task now runs the `desktop.ts` file.
  - The `dev` and `build` tasks now use Vite.
  - The `start` task runs the `main.ts` server for production.
